Refactor packaging: move under oidc
